renaming of several ops from feature to correct name
many ops were called feature, even if the classes are not part of the
feature package. therefore I renamed them to a more correct name.
