+++ src/main/java/net/imagej/ops/statistics/firstorder/DefaultMeanIRT.java
@@ -1,7 +1,4 @@
-package net.imagej.ops.statistics.firstorder.realtype;
-
 /*
-
  * #%L
  * ImageJ OPS: a framework for reusable algorithms.
  * %%
@@ -31,41 +28,47 @@ package net.imagej.ops.statistics.firstorder.realtype;
  * #L%
  */
 
+package net.imagej.ops.statistics.firstorder;
+
 import net.imagej.ops.AbstractOutputFunction;
 import net.imagej.ops.Op;
 import net.imagej.ops.OpUtils;
-import net.imagej.ops.features.firstorder.FirstOrderFeatures.SumFeature;
-import net.imagej.ops.statistics.FirstOrderOps.Sum;
-import net.imagej.ops.statistics.firstorder.FirstOrderStatIRTOps.SumIRT;
+import net.imagej.ops.features.firstorder.FirstOrderFeatures.MeanFeature;
+import net.imagej.ops.statistics.FirstOrderOps.Mean;
+import net.imagej.ops.statistics.firstorder.FirstOrderStatIRTOps.MeanIRT;
 import net.imglib2.type.numeric.RealType;
 import net.imglib2.type.numeric.real.DoubleType;
 
-import org.scijava.Priority;
 import org.scijava.plugin.Plugin;
 
 /**
- * Calculate {@link Sum} on {@link Iterable} of {@link RealType}
+ * Calculate {@link Mean} of {@link Iterable} of {@link RealType}.
  * 
  * @author Christian Dietz
- * @author Andreas Graumann
  */
-@Plugin(type = Op.class, name = Sum.NAME, label = Sum.LABEL, priority = Priority.LOW_PRIORITY)
-public class DefaultSumFeature<I extends RealType<I>, O extends RealType<O>>
-        extends AbstractOutputFunction<Iterable<I>, O> implements SumIRT<I, O>,
-        SumFeature<O> {
+@Plugin(type = Op.class, name = Mean.NAME, label = Mean.LABEL)
+public class DefaultMeanIRT<I extends RealType<I>, O extends RealType<O>>
+        extends AbstractOutputFunction<Iterable<I>, O> implements
+        MeanIRT<I, O>, MeanFeature<O> {
+
+    @Override
+    protected O safeCompute(Iterable<I> input, O output) {
+
+        double sum = 0;
+        double count = 0;
+
+        for (final RealType<?> val : getInput()) {
+            sum += val.getRealDouble();
+            ++count;
+        }
+
+        output.setReal((count == 0) ? 0 : sum / count);
+
+        return output;
+    }
 
     @Override
     public O createOutput(Iterable<I> in) {
         return OpUtils.<O> cast(new DoubleType());
-    }
-
-    @Override
-    protected O safeCompute(Iterable<I> input, O output) {
-        output.setReal(0);
-        for (final RealType<?> d : input) {
-            output.setReal(output.getRealDouble() + d.getRealDouble());
-        }	
-
-        return output;
     }
 }
