+++ src/main/java/net/imagej/ops/statistics/firstorder/DefaultSumOfLogsIRT.java
@@ -28,44 +28,44 @@
  * #L%
  */
 
-package net.imagej.ops.statistics.firstorder.realtype;
+package net.imagej.ops.statistics.firstorder;
 
 import net.imagej.ops.AbstractOutputFunction;
 import net.imagej.ops.Op;
 import net.imagej.ops.OpUtils;
-import net.imagej.ops.features.firstorder.FirstOrderFeatures.StdDeviationFeature;
-import net.imagej.ops.statistics.FirstOrderOps.StdDeviation;
+import net.imagej.ops.features.firstorder.FirstOrderFeatures.SumOfLogsFeature;
+import net.imagej.ops.statistics.FirstOrderOps.SumOfInverses;
+import net.imagej.ops.statistics.FirstOrderOps.SumOfLogs;
+import net.imagej.ops.statistics.firstorder.FirstOrderStatIRTOps.SumOfLogsIRT;
 import net.imglib2.type.numeric.RealType;
 import net.imglib2.type.numeric.real.DoubleType;
 
-import org.scijava.Priority;
 import org.scijava.plugin.Plugin;
 
-@Plugin(type = Op.class, name = StdDeviation.NAME, label = StdDeviation.LABEL, priority = Priority.LOW_PRIORITY + 1)
-public class DefaultStdDeviationFeature<I extends RealType<I>, O extends RealType<O>>
-		extends AbstractOutputFunction<Iterable<I>, O> implements StdDeviation,
-		StdDeviationFeature<O> {
+/**
+ * Calculate {@link SumOfInverses} on {@link Iterable} of {@link RealType}
+ * 
+ * @author Christian Dietz
+ * @author Andreas Graumann
+ */
+@Plugin(type = Op.class, name = SumOfLogs.NAME, label = SumOfLogs.LABEL)
+public class DefaultSumOfLogsIRT<T extends RealType<T>, O extends RealType<O>>
+        extends AbstractOutputFunction<Iterable<T>, O> implements
+        SumOfLogsIRT<T, O>, SumOfLogsFeature<O> {
 
-	@Override
-	public O createOutput(Iterable<I> in) {
+    @Override
+    public O createOutput(Iterable<T> in) {
         return OpUtils.<O> cast(new DoubleType());
-	}
+    }
 
-	@Override
-	protected O safeCompute(Iterable<I> input, O output) {
+    @Override
+    protected O safeCompute(Iterable<T> input, O output) {
 
-		double sum = 0;
-		double sumSqr = 0;
-		int n = 0;
-
-		for (final RealType<?> rt : input) {
-			final double px = rt.getRealDouble();
-			++n;
-			sum += px;
-			sumSqr += px * px;
-		}
-
-		output.setReal((Math.sqrt((sumSqr - (sum * sum / n)) / (n - 1))));
-		return output;
-	}
+        double result = 0.0;
+        for (final RealType<?> type : input) {
+            result += Math.log(type.getRealDouble());
+        }
+        output.setReal(result);
+        return output;
+    }
 }
